+++ rally/common/db/api.py
@@ -287,7 +287,7 @@ def task_update_status(session, uuid, status, allowed_statuses):
 
 
 @with_session
-def task_list(session, status=None, env=None, tags=None):
+def task_list(session, status=None, env=None, tags=None, uuids_only=False):
     tasks = []
     query = session.query(models.Task)
 
@@ -305,9 +305,13 @@ def task_list(session, status=None, env=None, tags=None):
             return []
         query = query.filter(models.Task.uuid.in_(uuids))
 
+    if uuids_only:
+        query = query.options(sa.orm.load_only("uuid"))
+
     for task in query.all():
         task = task.as_dict()
-        task["tags"] = sorted(tags_get(task["uuid"], consts.TagType.TASK))
+        if not uuids_only:
+            task["tags"] = sorted(tags_get(task["uuid"], consts.TagType.TASK))
         tasks.append(task)
 
     return tasks